requirements.txt
asgiref==3.11.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==26.1.0
colorama==0.4.6
crispy-bootstrap5==2026.3
decorator==5.2.1
Django==5.2.12
django-crispy-forms==2.6
django-debug-toolbar==6.2.0
django-environ==0.13.0
django-extensions==4.1
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.17.1
drf-spectacular==0.29.0
executing==2.2.1
factory_boy==3.3.3
Faker==40.12.0
inflection==0.5.1
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Markdown==3.10.2
matplotlib-inline==0.2.1
parso==0.8.6
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
wcwidth==0.6.0